refactor(cli): consolidate observability + inspection commands, remove check/resources [RUSH-1234]
Two overlapping command clusters had grown ambiguous. Give each surviving
command one clear role and delete the two genuinely redundant commands.

- Fold `agents check` (CI drift gate) into `agents doctor --check`: same
  `computeDrift` engine, now with a scriptable exit code. `--check --quiet`,
  `--check --json` (byte-compatible payload), and `--check --devices` (fleet
  gate) all carry over. The `--check` branch is placed before `--devices` so
  `doctor --check --devices` routes to the gate while `doctor --devices` still
  routes to the readiness matrix. Delete check.ts.
- Fold `agents resources` (merged first-wins cross-layer table) into
  `agents view --merged`. Lift the renderer into src/lib/merged-resources.ts
  (shared, so inspect can reuse it). Delete resources.ts.
- Remove both loaders + map entries from the command registry and index.ts
  eager registration; drop the check/resources lines from top-level help and
  point doctor/view at the new flags. The spellcheck registry auto-drops them.
- Document the observability taxonomy (events/feed/activity/output/sessions,
  one role each) in docs/06-observability.md, and doctor --check / view --merged
  across README, 00-concepts, and the changelog.

Verified: tsc clean; doctor.check + view.merged suites green; `doctor --check`
exits 1 on drift / 0 clean with the byte-compatible JSON shape; `view --merged`
renders the identical merged table; removed commands yield unknown-command.
